game phases update
